requirements.txt
altair==5.0.1
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.9
breadability==0.1.20
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
confection==0.0.4
cymem==2.0.7
decorator==5.1.1
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
filelock==3.12.1
fsspec==2023.6.0
gensim==3.8.3
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langcodes==3.3.0
lit==16.0.5.post0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.2
pathy==0.10.1
Pillow==9.5.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
plotly==5.15.0
preshed==3.0.8
protobuf==4.23.2
pyarrow==12.0.0
pycountry==22.3.5
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rich==13.4.1
safetensors==0.3.1
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-streamlit==1.0.6
srsly==2.4.6
streamlit==1.23.1
Summy==0.1
sumy==0.11.0
sympy==1.12
tenacity==8.2.2
thinc==8.1.10
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
transformers==4.30.1
triton==2.0.0
typer==0.7.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.3
validators==0.20.0
wasabi==1.1.2
watchdog==3.0.0
zipp==3.15.0
